Get rid of of historical conversion of m
In the beginning the script accepted n=7 and m=3, but the convention is
to call this a n=7 and m=4, because of the retrograde triangles. Later
the script accepted the parameters n=7 and m=4 (and not m=3) but
internally it still converted these to the old values. This is made more
consistent as preparation to handle odd m.

Also fixes a bug with 10 / 4 and 10 / 6 which had too many faces.
Otherwise I checked thall n/m for n in [5, 10] and m even generate the
correct polyhedra. TODO: make this an auto test.