roof tops: use '-s' and '-o' in file names
The extra triangles either end up at the same side or at the other side
of the base polygon compared to the triangles in the roof top.

Introduce 's' to express they are on the same side, 'o' for the opposite
side.

It is not practical to use the fold angle for this, since it is unknown
the dihedral angle in before folding.